split broadcast related stuff from apply.jl
