+++ src/TimeSeries.jl
@@ -21,6 +21,7 @@ include(".timeseriesrc.jl")
 include("timearray.jl")
 include("split.jl")
 include("apply.jl")
+include("broadcast.jl")
 include("combine.jl")
 include("readwrite.jl")
 include("utilities.jl")
